fix: integration tests

+++ app/router/routehomepage/controller_test.go
@@ -92,7 +92,7 @@ func controllerGetAllPosts(t *testing.T, ctrl *Controller) {
 			tt.args.ctx.SetParamValues(tt.args.paramValue)
 
 			if assert.NoError(t, ctrl.GetAllPosts(tt.args.ctx)) {
-				assert.Equal(t, tt.wantStatusCode, rec.Code)
+				assert.Equal(t, tt.wantStatusCode, int64(rec.Code))
 			}
 		})
 	}
